web-animations/animation-group/animation-group-last-child.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationGroup lastChild attribute test</title>
<meta name="assert" content="The last child of this animation group">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationgroup-lastchild">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var group = new ctor([]);

        assert_equals(group.lastChild, null, ctor.name + '.lastChild should return ' +
            'null if there are no children in the group');
    });
}, 'AnimationGroup.lastChild returns null if there are no children in the group');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node = newAnimation(createDiv(test));
        var group = new ctor([node]);

        assert_equals(group.lastChild, node, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
}, 'AnimationGroup.lastChild returns the last child of the animation group. ' +
    'Test the group with single child');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var group = new ctor([node1, node2]);

        assert_equals(group.lastChild, node2, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group. ' +
    'Test the group with several children nodes');

test(function() {
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = new AnimationGroup([]);
        var node2 = new AnimationGroup([]);
        var group = new ctor([node1, node2]);

        assert_equals(group.lastChild, node2, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group. ' +
    'Test AnimationGroup as a last child');

test(function() {
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = new AnimationSequence([]);
        var node2 = new AnimationSequence([]);
        var group = new ctor([node1, node2]);

        assert_equals(group.lastChild, node2, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group. ' +
    'Test AnimationSequence as a last child');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var node3 = newAnimation(createDiv(test));
        var node4 = new ctor([node1, node2]);
        var node5 = newAnimation(createDiv(test));
        var group = new ctor([node3, node4, node5]);

        assert_equals(group.lastChild, node5, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
        assert_equals(node4.lastChild, node2, ctor.name + '.lastChild should return ' +
            'the last child of low-level animation group');
    });
}, 'AnimationGroup.lastChild returns the last child of the animation group. ' +
    'Test tree structure');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var group = new ctor([node1, node2]);

        node1.before(node2);
        assert_equals(group.lastChild, node1, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is changed by method before()');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var node3 = newAnimation(createDiv(test));
        var group1 = new ctor([node1, node2]);
        var group2 = new ctor([node3]);

        node3.before(node2);
        assert_equals(group1.lastChild, node1, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is removed by method before()');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var group = new ctor([node1, node2]);

        node2.after(node1);
        assert_equals(group.lastChild, node1, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is changed by method after()');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var node3 = newAnimation(createDiv(test));
        var group1 = new ctor([node1, node2]);
        var group2 = new ctor([node3]);

        node3.after(node2);
        assert_equals(group1.lastChild, node1, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is removed by method after()');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var node3 = newAnimation(createDiv(test));
        var group = new ctor([node1, node2]);

        node2.replace(node3);
        assert_equals(group.lastChild, node3, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is changed by method replace()');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var node3 = newAnimation(createDiv(test));
        var group1 = new ctor([node1, node2]);
        var group2 = new ctor([node3]);

        node3.replace(node2);
        assert_equals(group1.lastChild, node1, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is removed by method replace()');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var group = new ctor([node1, node2]);

        node2.remove();
        assert_equals(group.lastChild, node1, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is changed by method remove()');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var group = new ctor([node1, node2]);

        group.prepend(node2);
        assert_equals(group.lastChild, node1, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is changed by method prepend()');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var node3 = newAnimation(createDiv(test));
        var group1 = new ctor([node1, node2]);
        var group2 = new ctor([node3]);

        group2.prepend(node2);
        assert_equals(group1.lastChild, node1, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is removed by method prepend()');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var group = new ctor([]);

        group.append(node1);
        assert_equals(group.lastChild, node1, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is changed by method append()');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = newAnimation(createDiv(test));
        var node2 = newAnimation(createDiv(test));
        var node3 = newAnimation(createDiv(test));
        var group1 = new ctor([node1, node2]);
        var group2 = new ctor([node3]);

        group2.append(node1);
        assert_equals(group1.lastChild, node1, ctor.name + '.lastChild should return ' +
            'the last child of the animation group');
    });
},  'AnimationGroup.lastChild returns the last child of the animation group, ' +
    'the last child is removed by method append()');
</script>
</body>
